Redesigned Configuration Code
The system should now use backup locations on failures.
If saving to the registry fails:
- Settings will be stored in an INI file in "%APPDATA%", or if that fails:
- Settings will be stored in an INI file in "%LOCALAPPDATA%", or if that fails:
- Settings will be stored in an INI file in "Documents", or if that fails:
- Settings will be stored in an INI file directly in your User directory.

The system uses a slightly different order if installed to "%LOCALAPPDATA%\Programs":
- Settings will be stored in a local INI file instead of "%APPDATA%"

The system also supports "portable" mode.
Instead of saving to the registry:
 - Settings will be saved in a Local INI file, or if that fails:
 - Settings will be saved in an INI file in the root of the Local Drive.

Each file has a `save` timestamp entry in the `[META]` group, which is used to identify the latest file.
All locations are scanned on startup, and the latest version is saved to the chosen location (if out-of-date).
Each save is sent to a temporary location, the original save is backed up, the temporary save is renamed, and the changes are verified before the backup is deleted, to prevent data loss.